restructure source code
