+++ app/service/shutdown_system_with_button.service
@@ -5,7 +5,7 @@ After=multi-user.target
 [Service]
 Type=simple
 Restart=always
-ExecStart=/usr/bin/python3 -u /home/pi/pixiebox/backend/pixiebox/shutdown_system_with_button.py
+ExecStart=/usr/bin/python3 -u /home/pi/pixiebox/app/service/shutdown_system_with_button.py
 StandardError=journal
 StandardOutput=journal
 StandardInput=null
