+++ app/service/tag/read_tag.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from common.rfid_reader import MFRC522Reader
+from lib.rfid_reader import MFRC522Reader
 
 rfidReader = MFRC522Reader()
 rfidReader.read()
