--- backend/pixiebox/common/tts.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python
-
-import pyttsx3
-
-
-class TTS:
-    @staticmethod
-    def say(text):
-        engine = pyttsx3.init()
-        engine.say(text)
-        engine.runAndWait()
