--- backend/pixiebox/install_complete.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-from common.tts import TTS
-
-tts = TTS()
-tts.say("Installation complete. Hello Pixiebox.")
