+++ scripts/install_complete_message.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+
+import pyttsx3
+
+engine = pyttsx3.init()
+engine.say("Installation complete. Hello Pixiebox.")
+engine.runAndWait()
